--- app/src/main/java/com/f2h/f2h_buyer/screens/group/daily_orders/DailyOrdersViewModelFactory.kt
@@ -1,18 +1,0 @@
-package com.f2h.f2h_buyer.screens.group.daily_orders
-
-import android.app.Application
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.f2h.f2h_buyer.database.SessionDatabaseDao
-
-class DailyOrdersViewModelFactory (
-    private val dataSource: SessionDatabaseDao,
-    private val application: Application) : ViewModelProvider.Factory {
-    @Suppress("unchecked_cast")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(DailyOrdersViewModel::class.java)) {
-            return DailyOrdersViewModel(dataSource, application) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
