+++ app/src/main/java/com/f2h/f2h_buyer/screens/group/group_tabs/GroupDetailsTabsAdapter.kt
@@ -3,26 +3,20 @@ package com.f2h.f2h_buyer.screens.group.group_tabs
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.f2h.f2h_buyer.screens.group.all_items.AllItemsFragment
-import com.f2h.f2h_buyer.screens.group.daily_orders.DailyOrdersFragment
 import com.f2h.f2h_buyer.screens.group.group_wallet.GroupWalletFragment
 
 
 class GroupDetailsTabsAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
 
-    override fun getItemCount(): Int = 3
+    override fun getItemCount(): Int = 2
     override fun createFragment(position: Int): Fragment {
         when(position) {
             0 -> {
-                val fragment = DailyOrdersFragment()
-                return fragment
-            }
-
-            1 -> {
                 val fragment = AllItemsFragment()
                 return fragment
             }
 
-            2 -> {
+            1 -> {
                 val fragment = GroupWalletFragment()
                 return fragment
             }
