+++ app/src/main/java/com/f2h/f2h_buyer/screens/group/group_tabs/GroupDetailsTabsFragment.kt
@@ -51,9 +51,8 @@ class GroupDetailsTabsFragment : Fragment() {
         val tabLayout = view.findViewById<TabLayout>(R.id.tab_layout)
         TabLayoutMediator(tabLayout, viewPager) { tab, position ->
             when(position) {
-                0 -> tab.text = "My Cart"
-                1 -> tab.text = "All Items"
-                2 -> tab.text = "Group Wallet"
+                0 -> tab.text = "All Items"
+                1 -> tab.text = "Group Wallet"
             }
         }.attach()
     }
