+++ app/src/main/java/com/f2h/f2h_buyer/screens/search_group/SearchGroupsViewModel.kt
@@ -109,7 +109,7 @@ class SearchGroupsViewModel(val database: SessionDatabaseDao, application: Appli
             searchGroupsItemList.add(uiElement)
         }
 
-        return searchGroupsItemList
+        return searchGroupsItemList.sortedBy { x -> x.groupName }
     }
