+++ app/src/main/res/menu/navdrawer_menu.xml
@@ -10,12 +10,15 @@
         android:icon="@drawable/settings_icon"
         android:title="@string/my_profile" />
     <item
-        android:id="@+id/signOutFragment"
-        android:icon="@drawable/sign_out_icon"
-        android:title="@string/sign_out" />
-    <item
         android:id="@+id/searchGroupsFragment"
         android:icon="@drawable/menu_icon"
         android:title="@string/search_groups" />
-
+    <item
+        android:id="@+id/dailyOrdersFragment"
+        android:icon="@drawable/my_cart"
+        android:title="@string/my_cart" />
+    <item
+        android:id="@+id/signOutFragment"
+        android:icon="@drawable/sign_out_icon"
+        android:title="@string/sign_out" />
 </menu>
