+++ app/src/main/res/values/strings.xml
@@ -10,6 +10,7 @@
     <string name="my_profile">My Profile</string>
     <string name="my_groups">My Groups</string>
     <string name="sign_out">Sign Out</string>
+    <string name="my_cart">My Cart</string>
     <string name="cancel">Cancel</string>
     <string name="save">Save</string>
     <string name="report">Order Reports</string>
